README.md
# Parque del Café - Integrative Task 3

This program reflects the functionality of the very well known Colombian park located in Quindio called El Parque Del Café.

## Tools:
* Java software.
* Eclipse.

## Participants:
* Daniela Olarte
* Santiago Gutierrez
* Esteban Mendoza

## Objective: 

The main objetive for the users who create their account is to be able to visualize the occupancy of the park in order to follow the Covid-19 protocols. While creating their account, users are also able to register their companions. After creating their account, users can choose the date of visit as long as the park has not yet reached 50% capacity. Users are able to purchase the different plans that the park offers for each companions and for itself. Users are able to see a brief description of the main games, food court and parking of the park as well as to register in each game/restaurant and to choose the parking of their preference. Finally, depending to the total amount paid, users can be eligible to win a benefit from the park for future visits. 

In addition, you can find all documentation of this game [here](https://github.com/danielaolarte123/TI2-snakes-and-ladders/tree/main/docs).